requirements.txt
clams-python==1.0.2
mmif-python==1.0.2
